Move SnapshotThreadLocal to common part
- to have the only implementation for all platforms
- remove unused ensureMutable
- use kotlin.native.concurrent.AtomicReference for AtomicReference implementation (instead of deprecated FreezableAtomicReference)
