+++ compose/runtime/runtime/src/commonMain/kotlin/androidx/compose/runtime/Expect.kt
@@ -19,16 +19,7 @@ package androidx.compose.runtime
 import androidx.compose.runtime.snapshots.Snapshot
 import androidx.compose.runtime.snapshots.SnapshotContextElement
 
-/**
- * This is similar to a `java.lang.ThreadLocal` but has lower overhead because it avoids a weak reference.
- * This should only be used when the writes are delimited by a try...finally call that will clean
- * up the reference such as [androidx.compose.runtime.snapshots.Snapshot.enter] else the reference
- * could get pinned by the thread local causing a leak.
- */
-internal expect class SnapshotThreadLocal<T>() {
-    fun get(): T?
-    fun set(value: T?)
-}
+internal expect fun getCurrentThreadId(): Long
 
 /**
  * Returns the hash code for the given object that is unique across all currently allocated objects.
@@ -61,8 +52,6 @@ internal expect class AtomicInt(value: Int) {
 }
 
 internal fun AtomicInt.postIncrement(): Int = add(1) - 1
-
-internal expect fun ensureMutable(it: Any)
 
 expect annotation class CompositionContextLocal
