+++ compose/runtime/runtime/src/jvmMain/kotlin/androidx/compose/runtime/ActualJvm.jvm.kt
@@ -25,22 +25,7 @@ import kotlinx.coroutines.ThreadContextElement
 
 internal actual typealias AtomicReference<V> = java.util.concurrent.atomic.AtomicReference<V>
 
-internal actual class SnapshotThreadLocal<T> {
-    private val map = AtomicReference<ThreadMap>(emptyThreadMap)
-    private val writeMutex = Any()
-
-    @Suppress("UNCHECKED_CAST")
-    actual fun get(): T? = map.get().get(Thread.currentThread().id) as T?
-
-    actual fun set(value: T?) {
-        val key = Thread.currentThread().id
-        synchronized(writeMutex) {
-            val current = map.get()
-            if (current.trySet(key, value)) return
-            map.set(current.newWith(key, value))
-        }
-    }
-}
+internal actual fun getCurrentThreadId(): Long = Thread.currentThread().id
 
 internal actual fun identityHashCode(instance: Any?): Int = System.identityHashCode(instance)
 
@@ -69,8 +54,6 @@ internal actual class AtomicInt actual constructor(value: Int) {
     actual fun set(value: Int) = delegate.set(value)
     actual fun add(amount: Int): Int = delegate.addAndGet(amount)
 }
-
-internal actual fun ensureMutable(it: Any) { /* NOTHING */ }
 
 /**
  * Implementation of [SnapshotContextElement] that enters a single given snapshot when updating
